Feat: Refactoring codebase to ts and adding pre-commit
