+++ .eslintrc.yml
@@ -1,0 +1,25 @@
+env:
+  browser: true
+  es2021: true
+  jest: true
+extends:
+  - eslint:recommended
+  - plugin:@typescript-eslint/recommended
+  - plugin:import/typescript
+  - plugin:prettier/recommended
+parser: '@typescript-eslint/parser'
+parserOptions:
+  ecmaVersion: latest
+  sourceType: module
+  project: ./tsconfig.json
+plugins:
+  - '@typescript-eslint'
+  - 'prettier'
+overrides:
+  - files:
+      - "src/**/*.ts"
+      - "src/*.ts"
+    rules:
+      '@typescript-eslint/no-explicit-any': ['error', { fixToUnknown: true }]
+ignorePatterns:
+  - src-js/
