+++ .gitignore
@@ -41,6 +41,7 @@ build/Release
 # Dependency directories
 node_modules/
 jspm_packages/
+src-js/
 
 # Snowpack dependency directory (https://snowpack.dev/)
 web_modules/
