+++ .prettierrc
@@ -1,0 +1,8 @@
+{
+    "tabWidth": 2,
+    "useTabs": false,
+    "semi": true,
+    "trailingComma": "all",
+    "singleQuote": false,
+    "printWidth": 80
+}
