--- src/index.js
@@ -1,12 +1,0 @@
-import { createServer } from './configs/server.config.js';
-import { connect } from './configs/db.config.js';
-import { DATABASE_URI } from './configs/constants.config.js';
-
-const app = createServer()
-
-const PORT = 5000;
-
-app.listen(PORT, async () => {
-    console.log(`Running on PORT ${PORT}`);
-    await connect(DATABASE_URI)
-})
